// src/service/request/type.ts
import type { AxiosRequestConfig, AxiosResponse } from 'axios'

export interface GpRequestInterceptors<T = AxiosResponse> {
  requestInterceptor?: (config: AxiosRequestConfig) => AxiosRequestConfig
  requestInterceptorCatch?: (error: any) => any
  responseInterceptor?: (config: T) => T
  responseInterceptorCatch?: (error: any) => any
}

export interface GpRequestConfig<T = AxiosResponse> extends AxiosRequestConfig {
  interceptors?: GpRequestInterceptors<T>
  showLoading?: boolean
}